requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
easydict==1.10
entrypoints==0.4
esphome==2022.2.6
fastjsonschema==2.16.2
fonttools==4.38.0
fvcore==0.1.5.post20221221
geomloss==0.2.5
idna==3.4
imageio==2.25.0
importlib-metadata==6.0.0
importlib-resources==5.10.2
iopath==0.1.10
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.5
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==2.6.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.6
open3d==0.10.0.0
packaging==23.0
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
pycparser==2.21
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytorch3d @ git+https://github.com/facebookresearch/pytorch3d.git@276c9a8acbfa715f5802f26ec9f4141bde26ecb4
PyWavelets==1.3.0
pywin32==305
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
tabulate==0.9.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
trimesh==3.18.3
typing_extensions==4.4.0
urllib3==1.26.14
vtk==9.2.5
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
wincertstore==0.2
yacs==0.1.8
zipp==3.12.1
